requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
python-dotenv==1.0.0
SQLAlchemy==2.0.21
gunicorn==21.2.0
pydantic==2.3.0
tiktoken==0.5.1
requests==2.31.0
openai>=1.0.0